pyproject.toml
[project]
name = "chatterbox-vllm"
description = "High-performance Chatterbox TTS on vLLM with OpenAI-compatible API, optimized for low VRAM GPUs"
dynamic = ["version"]
license = { file = "LICENSE" }
requires-python = ">=3.10"
readme = "README.md"
authors = [
    {name = "groxaxo"},
    {name = "David Li", email = "david@david-li.com"},
    {name = "resemble-ai", email = "engineering@resemble.ai"}
]
keywords = ["llm", "tts", "text-to-speech", "chatterbox", "vllm", "api", "openai", "multilingual", "low-vram"]
dependencies = [
    "torch",
    "torchaudio",
    "transformers",
    "tokenizers",
    "scipy",
    "numpy",
    "librosa",
    "s3tokenizer",
    "omegaconf",
    "conformer",
    "diffusers",
    "vllm",
    "peft",
    "llvmlite>=0.44.0",
    "vllm==0.10.0",
    "gradio",
    "spacy-pkuseg",
    "pykakasi",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
]

[project.optional-dependencies]
dev = ["build", "twine"]
ultra-low-vram = ["bitsandbytes>=0.41.0", "autoawq>=0.2.0"]

[tool.setuptools.dynamic]
version = { file = ".latest-version.generated.txt" }

[tool.setuptools.package-data]
chatterbox_vllm = ["models/t3/*.json", "models/*/t3/*.json"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.urls]
"Homepage" = "https://github.com/groxaxo/chatterbox-vllm2"
"Source" = "https://github.com/groxaxo/chatterbox-vllm2"
"Tracker" = "https://github.com/groxaxo/chatterbox-vllm2/issues"
"Documentation" = "https://github.com/groxaxo/chatterbox-vllm2/blob/master/README.md"
"Original vLLM Port" = "https://github.com/randombk/chatterbox-vllm"
"Original Chatterbox" = "https://github.com/resemble-ai/chatterbox"
